base commit

+++ app/Http/Controllers/Admin/Products/ProductController.php
@@ -124,7 +124,9 @@ class ProductController extends Controller
         if($product == null){
             return response()->json(['status' => false]);
         }
-        unlink(public_path('images/product/'.$product->image));
+        if($product->image != '' || $product->image != null){
+            unlink(public_path('images/product/'.$product->image));
+        }
         $product->delete();
 
         return response()->json(['status' => true]);
